parallel: Include output in CalledProcessError
The Parallel class captures all stdout/stderr for logging as each step
is completed. This change includes the output in any raised
CalledProcessError so that it can be used by Quibble elsewhere, e.g. for
transmitting error data to an external URL.

This change does not handle the scenario of forwarding subprocess output
for commands that are not invoked with Parallel.

Bug: T331061
Change-Id: Ife81df33a20fbf9b77223e2f1f3e5483799f9daf