commands: wrap subprocess in order to capture output
For the EarlyWarningBot we need to capture the commands output while
still emitted it to the console (for CI or humans).

This change is similar to the Unix `tee` command and unfortunately
Python does not provide the ability to both stream to stdout/stderr AND
capture to a buffer. The new `quibble.commands.run` implements a
capture, the process is opened in binary mode in order to display
potentially invalid unicode stream from the tests (see 5c29fb1b and
T318029).

The binary stream is read with `readline` which somehow
detects newline character and let us immediately output lines to the
user. The line is flushed immediately to preserve interativity when
a human watch an ongoing run.

The resulting output is decoded to unicode replacing any invalid unicode
with `backslashreplace` error handling, that can be safely handled as
text. A test is added to cover both handlings.

Tests have been changed to mock the higher level `quibble.commands.run`
instead of the low level `subprocess.Popen`.

Bug: T331061
Co-authored-by: Antoine Musso <hashar@free.fr>
Change-Id: I727c5f4f29219a47f131ae577ca6ef714270df52