Drop --run=all in favor of an empty list
Remove `--run` default value of `all` and internally use an empty list
when we want to check whether we want to run all stages (the default).

The reason is if we later on want to be able to pass multiple occurences
of `--run` we could use Python 3.8 `action='extends'` but that extends
the default (['all']). Thus passing `--run=phpunit --run=qunit` would
result in: ['all', 'phpunit', 'qunit']:

With the default 'store' action:

  >>> parser=argparse.ArgumentParser(); parser.add_argument('--run',default=['all'],type=comma_separated_list)
  >>> parser.parse_args(['--run', 'bar', '--run', 'foo'])
  Namespace(run=['foo'])

 With the 'extend' action:

  >>> parser.add_argument('--run',default=['all'],type=comma_separated_list,action="extend")
  >>> parser.parse_args(['--run', 'bar', '--run', 'foo'])
  Namespace(run=['all', 'bar', 'foo'])  # Default not overriden

With 'extend' action and without the default:

  >>> parser=argparse.ArgumentParser(); parser.add_argument(
  ...   '--run',type=comma_separated_list,action='extend')
  >>> Namespace(run=['bar', 'foo'])  # Correct

As a result of dropping `--run=all`, remove two unit tests:
* `--run=all` accepted all stages (we already test it is the default
  when run is not set)
* The one checking `--skip` applied to `--run=all` (we already test
  `--skip` works when no run is passed).

Change-Id: Ide2259739a0180e1bf7035f3172415c6c6ffff65