Run PHPUnit unit tests after installing MediaWiki
This is pretty much a revert of
I935004d8ed51d1c4a11c6db863ac182bd4d14b34.

While I would really like to say that unit tests don't need settings,
and to run them before installing MediaWiki, this utopic ideal clashes
with the current state of MW config, which is code running in the global
scope, and not data.

For the reasons explained in T227900#9014246, this step backwards is
necessary if we want to simplify our PHPUnit entry points.

Bug: T227900
Change-Id: Ie811aabecc67ce2d5b4310ea7b1e7ea664c10367