Print all deadlocks to MariaDB error log
When a build fails due to `Deadlock found when trying to get lock` we
are left with little indications about the root cause of the issue.

We had a long standing issue in MediaWiki with deadlock occuring when
concurrently creating users and the root cause was found by running
MariaDB with `innodb_print_all_deadlocks` enabled: T199393#4420174

Spinup the MariaDB server with `--innodb-print-all-deadlocks`, the
reports get printed to the error log which is written to the LOG_DIR as
of Ic0db2d4136575d0708853057dfe715952f47447b and thus end up attached to
the Jenkins build.

Bug: T342088
Change-Id: I003566eaf6db0ed41569d12a9a204a27c125a093