Remove MW_QUIBBLE_CI env variable
It couldn't be used reliably to detect Quibble because it was absent
during Apache requests for QUnit and Selenium.

The main case for which it was added, is the one case where the
MW_QUIBBLE_CI was intentionally absent, which is pure unit tests
where LocalSettings doesn't load. Those should not vary by environment.

If something needs to be turned on in Quibble, it can instead be
explicitly turned on by the way Quibble runs the thing that wants it.
For example, for T331621, quibble can pass a CLI option or dedicated
ENV variable about the option question. This way is remains
reproducible locally without needing to emulate Quibble fully, and
also avoids the env from being abused for other purposes.

Change-Id: I07fd2cb5a42eb9324f974c5ac7375effe67e8efd