Migrate from setup.py to pyproject.toml (PEP 517)
setup.py is deprecated in favor of PEP 517: "a build-system independent
format for source trees". Migrate the package metadata from `setup.cfg`
to `pyproject.toml`.

Debian Buster ships pip 18 which does not support the new build system,
upgrade pip.

Black infers the target python versions by reading `requires-python`
which is thus now set to Python 3.7+. Previously that was unset and
black would not enforce some styles such as always adding a trailing
comma in a list of method parameters (I supposed it was not supported in
Python 2.7). Fix the sole case in `tests/test_commands.py`.

Bug: T345093
Depends-On: Ib19f2a8b6800bed0744ccab19d089e33baa6e03a
Depends-On: If677e9a7d57f126c7ec9abd48cf9ed02dcce7f4f
Change-Id: I68686e668c3bac7cd1cb16c14a31f415d705edf9