Move MediaWiki install args to a standalone function
InstallMediaWiki.execute() is fairly large, move the logic to defines
install.php database options to a standalone "private" method.

Add fine unit tests to ensure we inject the proper set of options for
each of the supported databases and that we raise an exception when the
database type is unknown.

Mock the database options when testing execute() and simply ensure the
new helper method is invoked.

InstallMediaWikiTest no more inherits from unitttest.TestCase since we
used none of its method and that prevents pytest.mark.parametrize from
being applied.

Change-Id: Ia976a81e5e444e4841e0a2ac9b9679a7f199be13