Skip PostgreSQL test when it is not installed
I don't have PostgreSQL installed and whenever I run `tox` I end up with
a failure due to a missing pg_virtualenv. The test was written as a
quick way to verify we were able to spin up PostgreSQL and it is
working, so I guess we can ignore the failure.

Also make pytest verbose for integration tests to show the reason a test
got skipped.

Change-Id: Ib1934700db7e0ba47cc8ac1d5dc7b38f68a4b0e8