Remove unused util.php_version()
That was introduced by the change to support multiple workers in PHP
7.4+ web server: I7407466939ef3a1ff350b45f0f1920da23d9f361

At the time, we still supported php 7.2 and I guess some iteration of
the change wanted to only inject the setting when the PHP version being
used supports it. Since we set it via the environment variable
PHP_CLI_SERVERS_WORKERS, it does not matter whether the underlying php
supports it or not and php_version() has never been used.

Remove util.php_version() and its tests. As a side effect this removes a
dependency upon pkg_resources.Requirement which has been deprecated by
Python and triggers a DeprecationWarning when running tests.

Change-Id: I450231fc6c497df82a62e042ecd214483af7c0cb