Force git to fetch tags being updated
Since git 2.20, updates of tags are rejected by default:

  $ git fetch --tags
  From https://git.example.org/repo
   ! [rejected]        v1         -> v1  (would clobber existing tag)

It is a variation of 2020 bug T252310 which affected the Zuul merger
when some tag for pywikibot/core got changed. The analysis at
T252310#6222727 and the code are the same and we need git to force fetch.

The fix in the Zuul merger was If7eae5c0d703b7143c2d244f1c80391b2a1bf095

Bug: T356247
Change-Id: I44fe6a688555e09994df85f44b8feea3265d3433