Do not capture commands output in interactive mode
In 3c27b18cc we made the commands output to be captured which makes
stdin to not be attached and stdout/stderr are piped. When running
Quibble interactively (ex: quibble -c bash) the underlying commands
would test whether stdin/stdout are attached to a tty and since they are
not, the logic to bring interactivety get disabled. Essentially commands
are always run non interactively.

When stdin and stdout are attached to a tty, make our commands.run() to
invoke subprocess.check_call without touching stdout/stderr and return
early. This restores interactivity.

Bug: T360443
Change-Id: I53d0c6919df306f89381aad95984f8744bff74ac