Remove obsolete comment about --color
The Collapsing Console Sections Jenkins plugin had an issue with lines
starting with ANSI colors which I have fixed. At the same time, we made
Quibble to not output color unless stdout is a tty as a workaround which
is a good behavior to have (and on CI we can set `--color`).

Bug: T236222
Change-Id: Icb34620bb670b72b1c3e135302b9afbefa0cc6d6