Add parallel execution for PHPUnit extensions suite
Why:

The longest running CI jobs currently take more than 20 minutes to
complete, and the longest running part of those long jobs in the
PHPUnit `extensions` test suite. Some parts of the CI jobs (for
example the linting) already run in parallel, but the PHPUnit
tests run single-threaded on a single CPU.

Add functionality to Quibble to parallelise the execution of
PHPUnit test suites.

What:

The Mediawiki extensions test suite is a complicated suite of tests
that partially relies on dynamically generating PHPUnit test suites.
Splitting suites like this is currently not supported by extensions
like `paratest`. Instead, we need to make a list of all the test
classes that would be included in the suite (using PHPUnit's
`--list-test-xml`) and create "simple" suites composed of lists of
the PHP files defining the test classes.

Splitting the complete list of test classes over smaller buckets, we
can use Quibble and python multiprocessing to run multiple suites in
parallel and collect their output.

So that this is not enabled by default for all jobs, we make the
execution of parallel tests dependent on a new
`QUIBBLE_PHPUNIT_PARALLEL` environment variable / flag. The flag
should not be enabled for code-coverage jobs since the split runs
will overwrite each others coverage output.

The follow-up tasks T365978 and T365976 concern migrating the test
splitting to PHP so that developers can more easily reproduce these
runs locally.

Bug: T361190
Change-Id: I315b9aa1c86f998a7c44f8adb2c0a77ed5ec774e