Add support for core's composer `phpunit:...parallel:...` commands
Why:

Quibble 1.9.0 introduced parallel execution of PHPUnit tests.
The initial implementation was entirely in python in Quibble -
it did not require any changes to core code.

To make the same functionality available to developers on their
local setups, and to reduce the complexity of the Quibble
implementation, we introduce a composer command to run the tests
in parallel in If106802f08edd5d4c841bb7970c69b88ab3bb39b.

What:

Add commands to Quibble to make use of the new composer commands
for parallel PHPUnit tests. This will be the first step to
eventually deprecating the Quibble implementation of the same.

Bug: T365976
Change-Id: I447132aed3bca78c094dd69f7243c428b2e54f9b