Add notice at the end of console log for parallel test runs
The PHPUnit parallel test functionality is new and experimental,
and can lead to unexpected results from the CI jobs that have
parallel testing enabled.

Add a couple of log lines to the end of the parallel runs to give
developers information about where to look for support if they
encounter issues with parallel runs.

Bug: T361190
Change-Id: I81b6ef16824cb11fedff686964c379553c4227f6