Parallel PHPUnit: Only run Scributo SandboxTest if Scribunto loaded
For PHPUnit parallel testing, we have to make some exceptions for
Scribunto tests because they rely on dynamic generation of test
suites. Part of the special case logic excludes `SandboxTest.php`
during split group generation, and adds it manually back to the
first split group.

Not all Quibble test runs include the Scribunto extension. In this
case, blindly adding the `SandboxTest.php` back to `split_group_0`
causes the group to fail - the file is not present in the checkout.

We add a check here to ensure that the test is only added to the
split groups if the test file is present.

Bug: T368783
Change-Id: Ie5111a965521a3fc0016bd3f39823dd05abe4286