Fix "Multiple targets are not supported" with Chrome
When `CHROMIUM_FLAGS` is not set in the environment we inserted an empty
argument in the Chromium command line.

Since 110.0.5546.0, Chrome no more accept multiple URL to be given when
running in headless mode. chromium/chrome/app/chrome_main.cc has:

  if (headless::IsHeadlessMode()) {
    if (command_line->GetArgs().size() > 1) {
      LOG(ERROR) << "Multiple targets are not supported in headless mode.";
      return chrome::RESULT_CODE_UNSUPPORTED_PARAM;
    }
  }

Do not add the content of `CHROMIUM_FLAGS` when it is not trueish and
add a test to cover the issue. While at it, add a test to cover the
content of `CHROMIUM_FLAGS` environment variable is added to the list of
args.

That was reproducible locally by switching MediaWiki Qunit task from
Firefox to Chrome (reverting I6ecbd5743f3ed0d639b79cb08c8ff279e6ad9ff6 )
and using:

  DISPLAY='' CHROME_BIN=/usr/bin/chromium \
  quibble --skip-zuul --skip-deps --git-cache ~/projects --run=qunit

Ref: https://chromium-review.googlesource.com/c/chromium/src/+/4160268
Bug: T366799
Change-Id: Ieba6677537b7b1ed8249a450fe2080f46dd5b92c