Fix MW_SKIP_EXTERNAL_DEPENDENCIES to be set for is_vendor not use_vendor
Since I379ebe3fcd1ee99, we have been skipping composer.lock checks
on all jobs that involved mediawiki/vendor.

It was disabled for patches in *to* the mediawiki/vendor repo (as required)
but *also* for patches to all other repos in the one job that is able
to check the correctness of the mediawiki/vendor/composer.lock file,
which is the job that tests *with* mediawiki/vendor. (There is no need
to check composer.lock in jobs that install freshly from composer,
since composer generates the lock file during that job, we'd be
testing composer's ability to generate and immediately validate its
own lock file).

is_vendor is a subset of use_vendor,
e.g. on https://gerrit.wikimedia.org/r/c/mediawiki/vendor/+/1054291
the job "mediawiki-quibble-vendor-mysql-php74" runs with the
"--packages-source vendor" parameter passed to Quibble.

Bug: T370380
Change-Id: I87b0524ad4b4a49f0780e306ee910cef0f6de28f