Parallel PHPUnit: fix DatabaseLess result cache file name
The PhpUnitDatabaselessParallel class would invoke PHPUnit with the
result cache file set to:
.phpunit_group_{split_group_id}_databaseless.cache.json

That is because the string was incorrectly not a F-Strings. Since all
use of F-Strings were introduced recently as part of the Parallel
PHPUnit feature, remove all of those uses in favor of using percent
formatting as used in the rest of the code.

Add a test to cover PhpUnitDatabaselessParallel and
PhpUnitDatabaseParallel to ensure the testsuite name and the result
cache files are incremented between the first and last group.

Also remove F-Strings from quibble/phpunit_split (which is not covered
by tests).

Change-Id: Ic60c7a4ba1ce897c320b5d58f2f1959facfe6424