Experimental "success cache" support
As a final step in the execution plan, compute and store a cache entry
that represents a successful build/test of the distinct subject under
test, a SHA256 digest of:

* Data provided by the caller via `--success-cache-key-data`.
* The `HEAD^{tree}` hash of each of the sorted Git repos under the
  MediaWiki install directory.

Check for a success cache entry immediately following the cloning of all
projects by Zuul cloner. Note that for now, however, we will not skip
execution when encountering an entry. We are only experimenting to see
what the cache hit/miss rates would be.

The entries are stored in a new memcached based cache that can be
enabled in Quibble by passing `--memcached-server {server}`.

The cache is disabled by default and only enabled when both a memcached
server are provided and at least one `--success-cache-key-data` item is
given. The caller should determine what data represents the distinct
subject under test, e.g. job name, the container image ref/digest used
to execute Quibble, and any relevant job parameters.

Bug: T383243
Depends-On: Ifc6f8cc3c191d88300f8b2e007baf5d8fb0673e1
Change-Id: I0a436e64e09c0c810b67519a463fc00d01a00c86