Skip execution upon a success cache hit
The success cache appears to be working, averaging a 6% hit rate in
gate-and-submit and a significantly higher rate (~ 30%) for the
gate-and-submit-wmf and pipeline. The test-wmf pipeline has also shown a
higher hit rate at about 12%.

Move ahead with cache hit implementation by raising a new
`SuccessCache.Hit` exception and skipping all remaining commands when
catching that exception in `QuibbleCmd.execute`.

Note that it is typically an antipattern to use exceptions for flow
control. However, this particular pattern is similar enough to Python's
builtin `StopIteration`, so perhaps we need not fret about that.

Bug: T383243
Change-Id: Id2ed7a4f97e87178518451c49538135b576cb2ea