Fix logging exception when using --resolve-requires
The resolved dependencies are passed as a set to ZuulClone. Since
the introduction of the Chronometer the string representation of
ZuulClone is logged using a json.dumps() of its parameter. Since set are
not serializable to JSON, that raises an exception.

When converting the object to a str, have the JSON encoder to fallback
to a list.

Change-Id: I99d1cb615dcd90f1ca2e1bc06d7f65311d57fc79