Add support for OpenSearch
Some MediaWiki extensions interact with OpenSearch via the CirrusSearch
extension.

The code assumes:
- the server to be available on 127.0.0.1:9200 and would wait for the
  port to respond.
- the CirrusSearch MediaWiki extension to be available

After MediaWiki is installed, the code runs CirrusSearch maintenance
scripts UpdateSearchIndexConfig and ForceSearchIndex.

The feature can be enabled by setting `QUIBBLE_OPENSEARCH` environment
variable to a trueish value (1, true, yes, y...). On CI we set it to
either `false` or `true` which is then used in supervisor config to
determine whether OpenSearch is started or not.

Bug: T386691
Signed-off-by: Antoine Musso <hashar@free.fr>
Needed-By: I1b0961fa18c0e2ff06ff9f7c10925d91ce1aa10d
Change-Id: I931745bb8997bf64bbb1cefc8b21b6244119145b